Split offset memory from memory
